Set features context for both public and private API keys
